# packages/workflow-sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FlowAId API",
    "version": "1.1.0",
    "description": "Typed decision workflows: build, run, observe and evaluate. See docs/design/API.md."
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "`fa_live_…` / `fa_test_…` API key"
      },
      "session": { "type": "apiKey", "in": "cookie", "name": "__Host-fa_session" }
    },
...
  },
  "paths": {
...
    "/v1/workflows/{id}/versions": {
      "get": {
        "tags": ["versions"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "x-cli": { "noun": "version", "verb": "list", "positional": ["id"] },
        "x-auth": "session_or_api_key",
        "x-scopes": ["workflows:read"],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      },
                      "workflowId": {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      },
                      "kind": { "type": "string", "enum": ["published", "draft"] },
                      "version": { "type": ["number", "null"] },
                      "label": { "type": ["string", "null"] },
                      "definitionHash": { "type": "string" },
                      "planHash": { "type": "string" },
                      "compilerVersion": { "type": "string" },
                      "notes": { "type": ["string", "null"] },
                      "publishedBy": { "type": ["string", "null"] },
                      "createdAt": { "type": "string" }
                    },
                    "required": [
                      "id",
                      "workflowId",
                      "kind",
                      "version",
                      "label",
                      "definitionHash",
                      "planHash",
                      "compilerVersion",
                      "notes",
                      "publishedBy",
                      "createdAt"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/workflow-versions/{id}": {
      "get": {
        "tags": ["versions"],
        "parameters": [
          {
            "schema": { "type": "string", "enum": ["plan"] },
            "in": "query",
            "name": "include",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "x-cli": { "noun": "version", "verb": "get", "positional": ["id"] },
        "x-auth": "session_or_api_key",
        "x-scopes": ["workflows:read"],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "workflowId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "kind": { "type": "string", "enum": ["published", "draft"] },
                    "version": { "type": ["number", "null"] },
                    "label": { "type": ["string", "null"] },
                    "definitionHash": { "type": "string" },
                    "planHash": { "type": "string" },
                    "compilerVersion": { "type": "string" },
                    "notes": { "type": ["string", "null"] },
                    "publishedBy": { "type": ["string", "null"] },
                    "createdAt": { "type": "string" },
                    "definition": {},
                    "plan": {},
                    "diagnostics": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "code": { "type": "string" },
                          "severity": { "type": "string", "enum": ["error", "warning", "info"] },
                          "message": { "type": "string" }
                        },
                        "required": ["code", "severity", "message"],
                        "additionalProperties": {}
                      }
                    },
                    "catalogSnapshot": {
                      "type": "object",
                      "propertyNames": { "type": "string" },
                      "additionalProperties": { "type": "string" }
                    }
                  },
                  "required": [
                    "id",
                    "workflowId",
                    "kind",
                    "version",
                    "label",
                    "definitionHash",
                    "planHash",
                    "compilerVersion",
                    "notes",
                    "publishedBy",
                    "createdAt",
                    "definition",
                    "diagnostics",
                    "catalogSnapshot"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/workflow-versions/{id}/diff/{otherId}": {
      "get": {
        "tags": ["versions"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "otherId",
            "required": true
          }
        ],
        "x-cli": { "noun": "version", "verb": "diff", "positional": ["id", "otherId"] },
        "x-auth": "session_or_api_key",
        "x-scopes": ["workflows:read"],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/v1/workflow-versions/{id}/export": {
      "get": {
        "tags": ["versions"],
        "parameters": [
          {
            "schema": { "default": "json", "type": "string", "enum": ["json", "yaml", "ts"] },
            "in": "query",
            "name": "format",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "x-cli": { "noun": "version", "verb": "export", "positional": ["id"] },
        "x-auth": "session_or_api_key",
        "x-scopes": ["workflows:read"],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/v1/workflow-versions/{id}/restore-draft": {
      "post": {
        "tags": ["versions"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "x-cli": { "noun": "version", "verb": "restore", "positional": ["id"] },
        "x-auth": "session_or_api_key",
        "x-scopes": ["workflows:write"],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "draftRevision": { "type": "number" } },
                  "required": ["draftRevision"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/workflows/{id}/deployments": {
      "get": {
        "tags": ["deployments"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "x-cli": { "noun": "deployment", "verb": "list", "positional": ["id"] },
        "x-auth": "session_or_api_key",
        "x-scopes": ["workflows:read"],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      },
                      "environmentId": {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      },
                      "environment": { "type": "string" },
                      "versionId": {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      },
                      "version": { "type": ["number", "null"] },
                      "previousVersionId": { "type": ["string", "null"] },
                      "variableOverrides": {
                        "type": "object",
                        "propertyNames": { "type": "string" },
                        "additionalProperties": {}
                      },
                      "deployedAt": { "type": "string" },
                      "deployedBy": { "type": ["string", "null"] }
                    },
                    "required": [
                      "id",
                      "environmentId",
                      "environment",
                      "versionId",
                      "version",
                      "previousVersionId",
                      "variableOverrides",
                      "deployedAt",
                      "deployedBy"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/secrets/where-used": {
      "get": {
        "tags": ["secrets"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "query",
            "name": "credentialId",
            "required": true
          }
        ],
        "x-cli": { "noun": "secret", "verb": "where-used" },
        "x-auth": "session_or_api_key",
        "x-scopes": ["credentials:read"],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "workflowId": { "type": "string" },
                      "environmentId": { "type": "string" },
                      "secretName": { "type": "string" }
                    },
                    "required": ["workflowId", "environmentId", "secretName"],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [{ "apiKey": [] }, { "session": [] }]
}